backend/data_ingestion/requirements.txt
kafka-python==2.0.2
numpy==1.24.3
pandas==2.0.1
python-dotenv==1.0.0
faker==18.11.2
scipy==1.10.1
